Move validators to folder
