+++ src/Stratis.SmartContracts.Core.Validation.Tests/SmartContractDeterminismValidatorTests.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using Microsoft.CodeAnalysis;
 using Stratis.SmartContracts.Core.Validation.Validators;
 using Stratis.SmartContracts.Core.Validation.Validators.Module;
+using Stratis.SmartContracts.Core.Validation.Validators.Type;
 using Stratis.SmartContracts.Executor.Reflection;
 using Stratis.SmartContracts.Executor.Reflection.Compilation;
 using Xunit;
