+++ src/Stratis.SmartContracts.Core.Validation/Validators/Type/StaticConstructorValidator.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using Mono.Cecil;
 using Mono.Cecil.Rocks;
 
-namespace Stratis.SmartContracts.Core.Validation
+namespace Stratis.SmartContracts.Core.Validation.Validators.Type
 {
     /// <summary>
     /// Validates that a <see cref="Mono.Cecil.TypeDefinition"/> contains a static constructor
